Workaround for pw_dir/pw_shell fields returned by getpwnam/getpwuid
These functions always returns "/" as pw_dir and "/system/bin/sh" as pw_shell.
However, this is not convenient when we have own shell to run -
for example, bash. To properly fix that, we need to support kind of user database -
like in GNU/Linux (and other POSIX systems). Kind of /etc/passwd, /etc/shadow, etc.

However, even though it makes sense to do, it's not easy, and should be done
very accurate, to not break tons of existing software. So here is temporary
workaround - getpwnam/getpwuid looks on uid of requested user and, if it matches
current uid of the process, returns environment variables HOME and SHELL in
pw_dir and pw_shell fields. This is not that elegant, but at least it works.

TODO: implement proper user database and revert this workaround.
Signed-off-by: Dmitry Moskalchuk <dm@crystax.net>